# refactor: use regular blocking socket