--- a/src/rpc/socket.rs
+++ b/src/rpc/socket.rs
@@ -19,13 +19,6 @@ pub const DEFAULT_REQUEST_TIMEOUT: Duration = Duration::from_millis(2000); // 2 
 
 pub const READ_TIMEOUT: Duration = Duration::from_millis(50);
 
-// /// Minimum interval between polling udp socket, lower latency, higher cpu usage.
-// /// Useful for checking for expected responses.
-// pub const MIN_POLL_INTERVAL: Duration = Duration::from_micros(100);
-// /// Maximum interval between polling udp socket, higher latency, lower cpu usage.
-// /// Useful for waiting for incoming requests, and periodic node maintenance.
-// pub const MAX_POLL_INTERVAL: Duration = Duration::from_secs(1);
-
 /// Cleanup interval for expired inflight requests to avoid overhead on every recv
 const INFLIGHT_CLEANUP_INTERVAL: Duration = Duration::from_millis(200);
 
@@ -149,15 +142,6 @@ impl KrpcSocket {
 
         match self.socket.recv_from(&mut buf) {
             Ok((amt, SocketAddr::V4(from))) => {
-                // if self.poll_interval > MIN_POLL_INTERVAL {
-                //     if !self.inflight_requests.is_empty() {
-                //         self.poll_interval = MIN_POLL_INTERVAL;
-                //     } else if self.server_mode {
-                //         self.poll_interval = (self.poll_interval / 2).max(MIN_POLL_INTERVAL);
-                //     }
-                //     let _ = self.socket.set_read_timeout(Some(self.poll_interval));
-                // }
-
                 let bytes = &buf[..amt];
 
                 if from.port() == 0 {
@@ -222,12 +206,7 @@ impl KrpcSocket {
                 );
             }
             Err(error) => match error.kind() {
-                ErrorKind::WouldBlock => {
-                    // if self.poll_interval < MAX_POLL_INTERVAL {
-                    //     self.poll_interval = (self.poll_interval * 2).min(MAX_POLL_INTERVAL);
-                    //     let _ = self.socket.set_read_timeout(Some(self.poll_interval));
-                    // }
-                }
+                ErrorKind::WouldBlock => {}
                 _ => {
                     warn!("IO error {error}")
                 }