Compared current `main` branch with a simple blocking socket (50ms).

A simple blocking socket is consuming about the same CPU (+15%?) but is significantly faster (~1.2s).

```
Branch.  CPU ----- Avg. Time   Turns Sleep
main      1.36       5039ms     10        8s
PR         1.42        3593ms     10        8s
main      3.35       4595ms      30       8s
PR         3.68        3650ms.    30        8s
main      3.14        4850ms      30       1s
PR         3.25        3489ms     30        1s
```

Testing script

```rust
use std::{thread::sleep, time::{Duration, Instant}};
use mainline::Dht;

fn main() {
    let dht = Dht::client().unwrap();

    let mut elapsed_sum = 0;
    for i in 0..30 {
        let value = format!("teste{}", i);
        let value_bytes = value.as_bytes();
        println!("\n{i} - Storing immutable data {}", value);
        let elapsed = put_immutable(&dht, &value_bytes);
        elapsed_sum += elapsed;
        println!("Elapsed time: {:?} milliseconds", elapsed);
        println!("Average elapsed time: {:?} milliseconds", elapsed_sum / (i + 1));
        sleep(Duration::from_secs(1));
        println!()
    }

}

fn put_immutable(dht: &Dht, value: &[u8]) -> u128 {
    let start = Instant::now();
    dht.put_immutable(value).expect("put immutable failed");
    start.elapsed().as_millis()
}
```